+++ linux/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   media_kit_libs_linux
   screen_retriever_linux
   url_launcher_linux
+  webview_all_linux
   window_manager
 )